refactor(reconstruct): outsource the auto-selectors
